Silence warnings regarding possible loss of precision converting from integer to floating-point types. Removed unnecessary conversion to float in FTimecode::FromFrameNumber() Minimize loss of precision in FThreadHeartBeatClock::Seconds() for long-running titles FGenericPlatformMath::FRand() now truncates Rand() to 24 bits or RAND_MAX, whichever is smaller.  This prevents values closer to 1.0 being higher probability due to loss of precision. #rb luke.thatcher, eric.mcdaniel, brian.white #jira none
#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: jeff.newquist
#ROBOMERGE-SOURCE: CL 14574582 in //UE4/Release-4.26/... via CL 14574596 via CL 14575202
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v760-14685797)

[CL 14691361 by jeff newquist in Dev-Tools-Staging branch]

+++ Engine/Source/Runtime/Core/Private/HAL/LowLevelMemTracker.cpp
@@ -3616,7 +3616,7 @@ namespace LLMPrivate
 		auto WriteValue = [this, &TextBuffer](int64 Value)
 		{
 			TextBuffer.Reset();
-			TextBuffer.Appendf(TEXT("%0.2f,"), Value / 1024.0f / 1024.0f);
+			TextBuffer.Appendf(TEXT("%0.2f,"), (float)Value / 1024.0f / 1024.0f);
 			Write(TextBuffer);
 		};
 		auto WriteTag = [&WriteValue, &TagSizes, bTrackPeaks](const FTagData* TagData)
